+++ weatherly/__init__.py
@@ -11,7 +11,7 @@ __title__ = 'weatherly'
 __author__ = 'konradsic'
 __license__ = 'MIT'
 __copyright__ = 'Copyright 2023-present konradsic'
-__version__ = '0.4.0'
+__version__ = '0.5.0'
 
 from typing import NamedTuple, Literal
 
@@ -29,6 +29,6 @@ class VersionInfo(NamedTuple):
     micro: int
     release_type: Literal["alpha", "beta", "candidate", "final"]
     
-version_info = VersionInfo(major=0, minor=4, micro=0, release_type="final")
+version_info = VersionInfo(major=0, minor=5, micro=0, release_type="final")
 
 del NamedTuple, Literal, VersionInfo